backend/requirements.txt
fastapi
uvicorn
snowflake-connector-python
python-dotenv
pydantic
sqlalchemy
pydantic[email]
bcrypt
docx
PyPDF2
email-validator
jwt
